Update scripts for new run and evaluation modes

+++ jail/default_scripts/default_evaluate.sh
@@ -7,6 +7,7 @@
 
 # Load VPL environment vars.
 . common_script.sh
+
 if [ "$VPL_MAXTIME" = "" ] ; then
 	export VPL_MAXTIME=20
 fi
@@ -25,8 +26,9 @@ else
 	# Prepare the run script/program (vpl_execution)
 	./vpl_run.sh &>>vpl_compilation_error.txt
 	cat vpl_compilation_error.txt
-	if [ -f vpl_execution ] ; then
-		mv vpl_execution vpl_test
+	if [[ -f vpl_execution || -f vpl_wexecution ]] ; then
+		[ -f vpl_execution ] && mv vpl_execution vpl_test
+		[ -f vpl_wexecution ] && mv vpl_wexecution vpl_test
 		if [ -f vpl_evaluate.cases ] ; then
 			mv vpl_evaluate.cases evaluate.cases
 		else
@@ -46,6 +48,8 @@ else
 			cat vpl_environment.sh >> vpl_execution
 			echo "./.vpl_tester" >> vpl_execution
 		fi
+		chmod +x vpl_execution
+		apply_evaluation_mode
 	else
 		(
 			cat vpl_environment.sh
@@ -53,14 +57,10 @@ else
 			echo "echo"
 			echo "echo '<|--'"
 			echo 'echo "-$VPL_COMPILATIONFAILED"'
-			if [ -f vpl_wexecution ] ; then
-				echo "echo '======================'"
-				echo "echo 'It seems you are trying to test a program with a graphic user interface.'"
-			fi
 			echo "echo '--|>'"
 			echo "echo"
 			echo "echo 'Grade :=>>$VPL_GRADEMIN'"
 		) > vpl_execution
+		chmod +x vpl_execution
 	fi
-	chmod +x vpl_execution
 fi
